Add tests from autograder

+++ tests/correct3/class1.py
@@ -1,6 +1,6 @@
 # basic class init: tests basic class definition
 class Foo:
-  x::int = 6
+  x = 6
   def __init__(self):
     pass
